requirements.txt
Django>=5.0,<6.0
authlib>=1.3,<2.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
